backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pdfplumber==0.10.3
PyMuPDF==1.23.8
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.2.0
openai>=1.50.0
httpx>=0.27.0
spacy==3.7.2
python-dotenv==1.0.0
reportlab==4.0.9
aiofiles==23.2.1


